Update: Long IDs for Repos

+++ src/main/java/DrMuhamadMubarak/TheFuture/codeforge/repository/ProjectAnalyticsRepository.java
@@ -3,6 +3,6 @@ package DrMuhamadMubarak.TheFuture.codeforge.repository;
 import DrMuhamadMubarak.TheFuture.codeforge.model.ProjectAnalytics;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProjectAnalyticsRepository extends CrudRepository<ProjectAnalytics, Integer> {
+public interface ProjectAnalyticsRepository extends CrudRepository<ProjectAnalytics, Long> {
 
 }
